requirements.txt
# Core Dependencies
selenium>=4.15.0
webdriver-manager>=4.0.1
pandas>=2.0.0
openpyxl>=3.1.5

# Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.19.0
nltk>=3.9.1

# Data Processing
numpy>=1.26.0
scikit-learn>=1.7.0

# Image Processing
Pillow>=11.0.0

# Fuzzy Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.0

# Database (Optional)
psycopg2>=2.9.0
SQLAlchemy>=2.0.0

# Utilities
psutil>=5.9.0
python-dateutil>=2.9.0

# Development & Build
pyinstaller>=6.0.0

# Visualization (for notebooks)
matplotlib>=3.10.0
seaborn>=0.13.0
wordcloud>=1.9.0

# Jupyter (for notebook support)
jupyter>=1.1.0
ipykernel>=6.30.0
ipywidgets>=8.1.0
